requirements.txt
openai
gradio
pyttsx3
langdetect
google-cloud-texttospeech
python-dotenv